实验室管理系统-加入分页

+++ mgrsite/src/main/resources/static/lms/js/permission/permission.js
@@ -2,12 +2,12 @@ $(function () {
     //加载权限
     $(".reload-btn").click(function () {
         $.ajax({
-            url:'/reload',
-            type:'get',
-            success:function(data){
-                if(data.success){
+            url: '/reload',
+            type: 'get',
+            success: function (data) {
+                if (data.success) {
                     window.location.reload();
-                }else{
+                } else {
                     $.messager.popup(data.msg);
                 }
             }
@@ -15,20 +15,24 @@ $(function () {
     });
 
     //删除
-    $(".delete-btn").click(function(){
+    $(".delete-btn").click(function () {
         var id = $(this).parent().parent().children(":first-child").val();
-        if(id){
-            $.ajax({
-                url:'/permission/'+id,
-                type:'delete',
-                success:function(data){
-                    if(data.success){
-                        window.location.reload();
-                    }else{
-                        $.messager.popup(data.msg);
-                    }
+        bootbox.confirm("你确定要删除吗？", function (yes) {
+            if (yes) {
+                if (id) {
+                    $.ajax({
+                        url: '/permission/' + id,
+                        type: 'delete',
+                        success: function (data) {
+                            if (data.success) {
+                                window.location.reload();
+                            } else {
+                                $.messager.popup(data.msg);
+                            }
+                        }
+                    });
                 }
-            });
-        }
+            }
+        });
     });
 });
